Set Privapp Permissions Control to log
Even if it's already set on the ROM sources, the device tends to prioritize the device's props before the ROM ones. I've encountered that, by default, this prop is set on /vendor as "enforce". This commit sets it to log on both system and product to make sure it gets higher priority than the vendor one, making it set to log instead